Proxy floating IP calls to quantum
Fixes bug 1023169. Also fixes bug 1031119.

In Folsom, floating IP support is moved to Quantum. By this commit floating IP
calls to nova are proxied to Quantum and nova command can be used to manage
floating IPs.

The network flags defined in nova.network.manager are moved to a separate file.
nova.network.manager is not loaded when quantum is enabled but some flags are
also used in nova.network.quantumv2.api. In OpenStack config mechanism we cannot
define a flag with a same name in multiple places even if one of them are imported,
so I created a new file for the network flags. (No new flags are introduced)

Change-Id: I2c32948a8fe291601216dc62d546da64a8fb8428